requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
email-validator==2.1.0.post1

# Database
asyncpg==0.29.0
alembic==1.12.1
sqlalchemy[asyncio]==2.0.23

# Vector Database
qdrant-client==1.6.4

# LangChain & OpenAI
langchain==0.0.325
langgraph==0.0.10
langsmith==0.0.63
openai==1.3.5

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Frontend
streamlit==1.28.2

# HTTP Client
httpx==0.25.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# CLI Tools
typer==0.9.0
rich==13.7.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.0
